+++ src/cppboot/generate/context.py
@@ -14,10 +14,12 @@ class Context:
     namespace: str
     target: str
     macro: str
+    android_package: str
     project_dir: Path
     license_id: str
     with_modules: bool
     shared_library: bool
+    with_android_ci: bool
     with_vim: bool
     with_ctags: bool
     with_vscode: bool